Fixed consumer to manually ack messages, refactored rabbit utility code
